Now the user passes the paths of the rps directories and their types as command line arguments. The code parses argv, cretaes a base class pointer for each rps directory and calls the cropping function from the corresponding derived class

+++ main.cpp
@@ -8,38 +8,68 @@
 
 namespace fs = std::filesystem;
 
-int main() {
-	fs::path images_dir = fs::path(PROJECT_ROOT_DIR) / "F1";
+int main(int argc, char** argv) {
+	
 	fs::path cropped_save_dir = fs::path(PROJECT_ROOT_DIR) / "Cropped";
-	if (!fs::exists(images_dir)) {
-		std::cerr << "Error: Directory 'F1' does not exist." << std::endl;
-		return 1;
-	}
 
+	for (int i = 0; i < argc; i++) {
+		std::string arg = argv[i];
+		
+		if (arg == "--rps") {
+			if (i + 2 >= argc) {
+				throw std::runtime_error("Error: --rps option must have 2 arguments: 'path' and 'type'");
+			}
+
+			std::string path;
+			std::string type;
+			if (i + 2 < argc) {
+				path = argv[i + 1];
+				type = argv[i + 2];
+			}
+			if (type != "simple" && type != "complex") {
+				throw std::runtime_error("Error: Invalid image type: " + type + ". The type should be either 'simple' or 'complex'");
+			}
+
+			std::cout << "Processing images in directory: " << path << " with type: " << type << std::endl;
+			fs::path images_dir = fs::path(PROJECT_ROOT_DIR) / path;
+			if (!fs::exists(images_dir)) {
+				throw std::runtime_error("Error: Directory" + path + " does not exist at " + images_dir.string());
+			}
+
+			if (fs::exists(cropped_save_dir)) {
+				fs::remove_all("Cropped");
+			}
+
+			fs::create_directories(cropped_save_dir);
+			if (!fs::exists(cropped_save_dir)) {
+				throw std::runtime_error("Error: Could not create directory 'Cropped' at " + cropped_save_dir.string());
+			}
+
+			std::string cropped_path = cropped_save_dir.string();
+
+			for (const auto& entry : fs::directory_iterator(images_dir)) {
+				std::string img_path = entry.path().string();
+				if (!fs::exists(img_path)) {
+					throw std::runtime_error("Error: Image file does not exist at " + img_path);
+				}
+
+				ImageCropper* cropper;
+				if (type == "simple") {
+					cropper = new SimpleImageCropper();
+				}
+				else {
+					cropper = new ComplexImageCropper();
+				}
+				cropper->img_crop(img_path, cropped_path);
+			}
+			i += 2;
+		}
+	}
+	
+	fs::remove_all(cropped_save_dir);
 	if (fs::exists(cropped_save_dir)) {
-		fs::remove_all("Cropped");
-	}
-
-	fs::create_directories(cropped_save_dir);
-	if (!fs::exists(cropped_save_dir)) {
-		std::cerr << "Error: Could not create directory 'Cropped'." << std::endl;
+		std::cerr << "Error: Could not remove directory 'Cropped'." << std::endl;
 		return 1;
 	}
 
-	std::string cropped_path = cropped_save_dir.string();
-
-	for (const auto& entry : fs::directory_iterator(images_dir)) {
-		std::string img_path = entry.path().string();
-		if (!fs::exists(img_path)) {
-			std::cerr << "File does not exist: " << img_path << std::endl;
-		}
-		img_crop(img_path, cropped_path);
-
-		fs::remove_all(cropped_save_dir);
-		if (fs::exists(cropped_save_dir)) {
-			std::cerr << "Error: Could not remove directory 'Cropped'." << std::endl;
-			return 1;
-		}
-		return 0;
-	}
 }
